+++ src/NosCore.Core/HttpClients/IncommingMailHttpClients/IncommingMailHttpClient.cs
@@ -38,7 +38,7 @@ namespace NosCore.Core.HttpClients.IncommingMailHttpClients
             RequireConnection = true;
         }
 
-        public async Task DeleteIncommingMailAsync(int channelId, long id, short mailId, byte postType)
+        public async Task DeleteIncommingMailAsync(long channelId, long id, short mailId, byte postType)
         {
             using var client = await ConnectAsync(channelId).ConfigureAwait(false);
             if (client == null)
@@ -48,7 +48,7 @@ namespace NosCore.Core.HttpClients.IncommingMailHttpClients
             await client.DeleteAsync(new Uri($"{client.BaseAddress}{ApiUrl}?id={id}&mailId={mailId}&postType={postType}")).ConfigureAwait(false);
         }
 
-        public async Task NotifyIncommingMailAsync(int channelId, MailData mailRequest)
+        public async Task NotifyIncommingMailAsync(long channelId, MailData mailRequest)
         {
             using var client = await ConnectAsync(channelId).ConfigureAwait(false);
             if (client == null)
@@ -60,7 +60,7 @@ namespace NosCore.Core.HttpClients.IncommingMailHttpClients
             await client.PostAsync(new Uri($"{client.BaseAddress}{ApiUrl}"), content).ConfigureAwait(false);
         }
 
-        public Task OpenIncommingMailAsync(int channelId, MailData mailData)
+        public Task OpenIncommingMailAsync(long channelId, MailData mailData)
         {
             throw new NotImplementedException();
         }
